requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiolimiter==1.1.0
aiosignal==1.3.2
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.95.0
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.5.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-api-core==2.24.2
google-api-python-client==2.178.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imbalanced-learn==0.12.4
importlib_resources==6.5.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.4.2
kiwisolver==1.4.7
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
matplotlib==3.9.4
multidict==6.4.3
narwhals==1.34.1
numpy==2.0.2
oauthlib==3.3.1
openai==1.86.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
plotly==6.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.21
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
reportlab==4.4.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.31.0
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.7
starlette==0.26.1
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.21.1
websocket-client==1.8.0
wsproto==1.2.0
xgboost==2.1.4
yarl==1.19.0
zipp==3.21.0
